* kernel/devices/keyboard/devkbmap.c: unicode pb maybe syncweb issue?

+++ kernel/devices/portfns_devices.h
@@ -23,7 +23,7 @@ void    drawcmap(void);
 // mouse/devmouse.c
 void    mouseresize(void);
 // keyboard/portkbd.c
-void kbdputsc(int, int);
+void kbdputsc(byte, int);
 int   kbdgetmap(uint, int*, int*, Rune*);
 void    kbdputmap(ushort, ushort, Rune);
